requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tweepy==3.8.0
urllib3==1.25.9
websocket-client==0.48.0
ibm-cos-sdk==2.6.3
ibm-cos-sdk-core==2.6.3
ibm-cos-sdk-s3transfer==2.6.3